Create Repositories to handle get requests
Create an implmentation of CrudRepository for each of
catalogs, components, profiles, and ssps to reduce
code duplication and set up for future CRUD and
persistence operations. Corresponding OSCAL objects
were created as well for this purpose.

Add properties to application.properties to allow
for easy use of the CRUD repositories and handling
of HTTP requests.

Added an exception for use in repostories when
handling HTTP requests.