Add tests and file existence checks
Add repository tests to test that the repository
methods return the correct value or throw the
correct exception. These tests are a basis for
how the repository code will be tested in the future.

Add checks in OscalRepository for file existence or if
the provided id is a directory to avoid returning
Optional.empty for all IOExceptions.