Clean Up GitHub Actions Config
Make cleanup changes to pom.xml and gh-packages.yml

Trigger on push to `develop` instead of `main`, as no code is begin
pushed to main for now.
Use the supported `temurin` distribution instead of `adopt` in the setup
java action.
Add caching of maven package dependencies to setup java action.
Fix a typo.