Update Packages Workflow for Repository Separation
Updating the GitHub Packages Workflow to publish only the
`oscal-rest-service-app` module of the REST service, because it is
the module that needs to be downloaded by the all-in-one deployment
Docker container.
This change is necessary after the separation of the repository
into multiple modules; now only one package will be created.