Download `oscal-demo-content` for test content
This downloads the file as part of the maven build process, though none
of the files are actually used yet. Redirects are allowed so that the
zip can be downloaded successfully. The ZIP is unpacked to the directory
where the resources previously were placed during the build.

The name of the resulting directory ends up being a bit ugly. This is an
artifact of how GitHub ZIP downloads work. It is
`<repoName>-<branchName>`, so we can `oscal-demo-content-main` as the
exploded directory name in `target/classes/`.