Leverage the `oscal-demo-content` artifacts
This actually uses the data across the application. Despite the name
being uglier, that should not impact usage of the REST service in "real
world" use cases (such as the `all-in-one` image or even standalone
deployments) since it really ought to be overridden from
`target/<something>` anyway and end users are not required to use this
naming convention.