Properly set the Content-Type for resources
This ensures that we make a best-effort attempt to set the Content-Type
header for response objects (at least, a little better than running with
Octet Stream for everything). Doing so requires using the
`probeContentType` method on `Files`. If that fails, then we still fall
back to an Octet Stream which is good for ensuring the output is treated
as binary data (which resources almost always are).

The tests also were not running for this route because the test class
was `abstract`. We now use a `StreamingResponseBody` for this route as
well for consistency in tests and implementation. Also, not doing that
resulted in other tests randomly failing. We can just use `Files.copy`
for this so that we don't have to deal with performing the transfer
ourselves manually.