Update to liboscal-java 1.0.4
This performs the basic part of the update to adopt the new version;
however, we're going to need to decide how to handle the validation
warnings that come as part of this update. In some scenarios, they seem
to be rather strict so we likely need to look into that further.