Upgrade to Spring Boot 3 (#155)
This makes a few significant changes:

- Upgrades minimum required Java version to 17 (latest LTS)
- Removes Springfox and the Swagger UI
- Upgrages Spring Boot and related tooling to the latest version

This also results in the ability to remove a few steps from `pom.xml`
and change the GitHub workflows. The README is also updated to point to
other documentation for the REST API specification.

We may try to find an alternative to SpringFox at some point to allow
us to document the API (and especially focus on the parts we implement);
however, relying on the shared documentation seems sufficient for now.