feat(commons): add a "null" constraint validation handler
This is an attempt to use the `IConstraintValidationHandler` plumbing in
liboscal-java to have a handler that does nothing at all. At the moment
for some reason, despite using the handler, it seems that logs are still
emitted on constraint validation events. I am curious about the cause
there.

This **is not** intended for actual use or to be merged. This is pretty
much purely an academic exercise just to create a minimal test and see
what happens.