chore(pipeline): add Java 20 to test matrix
Java 20 is expected some time within the next month. We need to remember
to add this to our testing matrix in the pipeline. Technically, we claim
to only test the "latest" stable release, but we'll go ahead and
continue testing two (we need Java 17 for LTS still anyway).

This needs to not be merged until Java 20 is GA. It is currently in the
RC phase. It is expected on March 21.