chore(deps): Bump spring-boot-starter to 3.0.4 (#178)
This handles the bump for
`spring-boot-starter-parent` from 3.0.2 to 3.0.4.
It appears we do not need to define a version
for `spring-boot-starter-data-jpa`, unless we
desired a version differing from the starter
parent. Maven by default will download
jar files based on the version defined for the
starter parent tag, which will be used as the
version towards dependencies with the
`org.springframework.boot` `groupId`. Occurrences of
`spring-boot-starter-data-jpa.version` have been
removed to avoid additional spring boot starter
PRs.