feat(commons): add a "no-op" constraint validation handler (#182)
This is an attempt to use the `IConstraintValidationHandler` plumbing in
liboscal-java to have a handler that does nothing at all. This hopefully
fixes
some issues that we are having where logging (which is the default
handler) is
in the hot path and the response doesn't get returned until the logs are
emitted.
And for some reason, there are a lot of validation failures so these
logs take
quite awhile to be written. This should significantly cut down the
amount of time
it takes to return a response to the `GET /catalogs` and `GET
/catalogs/{uuid}`
endpoints.

If this isn't fast enough, we can just (temporarily) disable constraint
validation
until we can see what's up with these duplicate key constraint
violations.